Bug 663597 - Memory leaks in normalization code
The DN normalization code uses a Slapi_Attr on the stack to avoid
allocation of the struct.  The contents of the Slapi_Attr are
never freed.  This patch ensure that the struct is cleared out
properly.

There was also a leak in the syntax normalization code where a
pointer to recently allocated string could get overwritten without
freeing the string first.  This patch frees the string first.