Bug 663484 - Entry usn plugin fails to properly tag entries on initialization
https://bugzilla.redhat.com/show_bug.cgi?id=663484

Description:
1) Introducing a config parameter nsslapd-entryusn-import-initval.
If the value is digit, e.g., 0, 10, 100 ..., the imported entries
will have the entryusn value.
If the value is not digit, e.g., "next", the imported entries
will have the last entryusn + 1 from the database existed before
the import was exected or initialized by the master.

2) Internal OID LDBM_ENTRYUSN_OID has been redifined in back-ldbm.h
to avoid a possible conflict.

3) If an attribute to modify is an operational attribute and
that is the only one modify operation (not one of the series
of modify operations), the server returns UNWILLING_TO_PERFORM.

In addition, this patch fixes a minor invalid access to a freed
memory in cl5DeleteDBSync.