Bug 661102 - Rename of managed entries not handled correctly
When usign the managed entry plug-in, a modify operation against
an origin entry that should trigger a rename of a managed entry
does not trigger the rename.  In addition, an error message is
logged since we try to do a modify of the RDN attribute of the
managed entry.

This patch makes the modify callback check if the changes require
a rename of the managed entry when an origin entry is updated.  If
a rename is needed, we perform a MODRDN.  We also need to skip the
RDN attribute when generating the mapped modifications that we use
to update a managed entry.