Bug 667488 - Cannot recreate numsubordinates index with db2index
https://bugzilla.redhat.com/show_bug.cgi?id=667488

Description: Special thanks to Orion Poplawski(orion@cora.nwra.com)
for his in-depth analysis on the bug.  As he figured out, a helper
function add_op_attrs called from the reindex API ldbm_back_ldbm2index
removes the numsubordinates attribute-value pair from entries to
be reindexed, which makes reindexing numsubordinates fail.

The helper function add_op_attrs adds operational attributes parentid,
entryid and entrydn to entries as well as removes numsubordinates
from them before the entries are reindexed/imported. The add and
remove are necessary for the import, but not for the reindex since
these to-be-added operational attributes are already in entries
in the primary database id2entry.  And numsubordinates is necessary
when the attribute is reindexed.

This patch updates ldbm_back_ldbm2index (ldif2ldbm.c) not to call
add_op_attrs.