Bug 201652 - LDAPv2 bind with expired password doesn't unbind correctly
When an LDAPv2 client binds with an expired password, we call the
unbind code to disconnect the client.  The unbind code doesn't handle
this well since we don't have an actual unbind operation to parse.

It seems wrong to call the unbind code to disconnect the client, as
we don't have an actual unbind operation to process and we don't
want to call any unbind plug-ins.  The proper thing would be to
just disconnect the client without going through all of the unbind
operation processing.