Bug 663752 - Cert renewal for attrcrypt and encchangelog
https://bugzilla.redhat.com/show_bug.cgi?id=663752

Description: In fixing Bug 182507, the feature to encrypt changelogs
had been introduced. The changelog encryption depends on the server
certificate as the attrcrypt does.  When the server certificate is
renewed, the encrypted changelog won't be decrypted.  This patch
implements/completes the feature to export and import the contents
of the changelog to provide the scenario to update the encrypted
changelogs along with the cert renewal.

See also this section for the steps to export/import changelogs.
http://directory.fedoraproject.org/wiki/Changelog_Encryption#Steps_for_Certificate_Renewal