Bug 470576 - Migration could do addition checks before commiting actions
https://bugzilla.redhat.com/show_bug.cgi?id=470576

The migration script can fail for various reasons leaving the server
in an inconsistent state. Sometimes a manual cleanup is required before
the script can be run again. Since there is no easy way to revert the
changes done to the server, the script has been modified such that it
removes the server that fails to be migrated. Once the cause of the
failure has been resolved the script can be run again.