Bug 481195 - Missing op type in log when password change required
When a password change is required, non-password change operations
are rejected with err=53.  The access log currently doesn't list
anything about the operation type in the access log.  This is
somewhat by design as we want to do as little processing as possible
for rejected operations.  To be consistent with other code that
does an early reject of operations (such as siabling anonymous
access or SSF restrictions), we should explicitly log these as an
"UNPROCESSED OPERATION".