Bug 669205 - db2bak: backed up changelog should include RUVs
https://bugzilla.redhat.com/show_bug.cgi?id=669205

Description:
Introduced backup plugin hooks: SLAPI_PLUGIN_BE_PRE_BACKUP_FN
and SLAPI_PLUGIN_BE_POST_BACKUP_FN to call back cl5WriteRUV and
cl5DeleteRUV, respectively.  cl5WriteRUV adds RUVs to changelog
and cl5DeleteRUV reads and deletes RUVs in changelog.  The call-
back functions are avaiable only when the process is initialized
as a server, which must have started with a backend normal mode
flag (DBLAYER_NORMAL_MODE) not with other utility modes such as
DBLAYER_ARCHIVE_MODE.  With this restriction, db2bak is not
allowed to use to back up the database including changelog db
when the server is up.  If launched, the utility fails with this
error message:
  [...] - db2archive: pre-backup-plugin failed (1).
  [...] - ERROR: Standalone db2bak is not supported \
  when a multimaster replication enabled server is coexisting.
  Please use db2bak.pl, instead.
As mentioned in the message, db2bak.pl is supposed to be used.

See also:
http://directory.fedoraproject.org/wiki/Move_changelog#Backing_up_Changelog