Bug 671033 - range sharing between server breaks with SASL/GSSAPI auth
The DNA range request extop uses the replication bind credentials
to authenticate to the other master.  If the replication agreement
is configured to use SASL/GSSAPI (or any non simple bind auth), the
DNA plug-in falls back to using an anonymous bind, which will fail.

The problem is that the function that fetches the replication bind
credentials was expecting a bind dn and password.  If both of these
were not found, it would free everything (including the bind
method).  The fix is to just fetch the credentials and return them
without doing this additional check.

This patch also makes the receiving master check if the bind DN
matches any replication bind DN, which allows it to work properly
when multiple replication bind DNs are configured.  It previously
only worked with a single-valued replication bind DN attribute.