Bug 527912 - setup-ds.pl appears to hang when DNS is unreachable
The setup-ds.pl program will appear to hang when it attempts to
contact an unreachable DNS server to determine the FQDN.  The
hostname prompt will eventually be displayed after a network
timeout is hit.

This patch makes the setup program log a warning message just
prior to trying to determine the FQDN.  This message states that
setup may appear to hang if it can't reach the DNS servers, and
that one can re-run setup and pass the hostname as an option on
the command line instead.