Bug 576534 - Password displayed on console when entered in command-line utilities
https://bugzilla.redhat.com/show_bug.cgi?id=576534
Resolves: bug 576534
Bug Description: Password displayed on console when entered in command-line utilities
Reviewed by: nhosoi (Thanks!)
Branch: master
Fix Description: Add a new configurable path - sttyexec - to configure.ac.
This is the absolute path and filename of the stty command to use with
the -echo and echo options to disable and enable tty echo for password
entry with perl scripts.  By default it is set to /bin/stty but it can be
overridden on a per-platform basis in configure.ac.  I had to move
DialogManager.pm to DialogManager.pm.in in order to replace the stty
command used there (which actually worked with just stty - not sure
why that worked but other perl scripts did not).
Platforms tested: RHEL6 x86_64
Flag Day: yes - file renamed - autoconf file changes
Doc impact: no