Bug 671199 - Don't allow other to write to rundir
The persmissions on /var/run/dirsrv currently get set to 777 by
the setup program.  There were some discrepencies with the way the
changeOwnerMode subroutine is used  that cause 777 to be the mode
set when we intended for it to be 770.  This patch fixes up the
way changeOwnerMode is used to sllow one to reset the group
ownership without altering the permissions for other.

In addition, this patch makes an upgrade remove any permissions that
are set for other on the rundir.