Bug 668619 - slapd stops responding
https://bugzilla.redhat.com/show_bug.cgi?id=668619

Description: The server uses c_timelimit field in the connection
table element to check the staled Simple Paged Results request.
The field is supposed to be cleaned up when the staled Simple
Paged Results connection is disconnected.  But the cleanup was
sometimes incomplete.  It causes the following requests which
happens to acquire the same connection table element timed out.
This patch forces to clean up the c_timelimit every time the
connection table is acquired and renewed.  Also, the timeout
check is done only for the Simple Paged Results connection.