Bug 667935 - DS pipe log script's logregex.py plugin is not redirecting the log output to the text file
https://bugzilla.redhat.com/show_bug.cgi?id=667935
Resolves: bug 667935
Bug Description: DS pipe log script's logregex.py plugin is not redirecting the log output to the text file
Reviewed by: nkinder (Thanks!)
Branch: master
Fix Description: The logregex plugin function is really just a thin wrapper
around the builtin plugin - the only difference is that logregex will only
store those lines that match the given regexes.  The fix is to just remove
the post function, and make the plugin just call the main plugin function
with the line if it matches the regexes.  The only tricky part is that we
have to increment total lines in the logregex plugin and decrement it when
we call the main pluginfunction, in order to keep an accurate count of the
total number of lines read.  I also cleaned up an error message and made
the print out flush.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no