Bug 670616 - Allow SSF to be set for local (ldapi) connections
This patch adds a new config parameter named nsslapd-localssf.
This parameter can be set to the SSF that one wants to apply for
ocal (LDAPI) connections.  This SSF will be used with the minssf
global and ACI settings.  The local SSF can also be used to satisfy
the confidentiality requirements for secure binds and password
modify extended operations.