Bug 668862 - init scripts return wrong error code
The dirsrv init script returns an exit code of 0 when no instances
are configured or if an invalid instance name is specified.  This
patch makes the dirsrv init script return the proper exit codes.
The exit codes for the status action are different than the codes
for non-status actions per the Fedora SysV init script packaging
guidelines.

In addition, the dirsrv and dirsrv-snmp inistscripts need to return
a non-0 exit code when networking is disabled or if the binaries
do not exist.  This patch handles those cases as well.