Bug 674430 - Improve error messages for attribute uniqueness
The attribute uniqueness plug-in always returns the same error
message when it encounters a problem.  The error message indicates
that there is a uniqueness conflict even when that is not the case.

This patch makes the error message only report a uniqueness conflict
when there really is a uniqueness conflict.  All other errors will
report that there was a problem checking for uniqueness.  This patch
also makes a MODRN return an error 32 when the target does not exist
instead of returning an operations error.