Bug 675853 - dirsrv crash segfault in need_new_pw()
A bind DN with a trailing tab in the RDN can cause ns-slapd to crash.
This is ultimately caused by the function that explodes the bind DN
trimming off the tab character, causing the bind entry to be found
by the backend.  The frontend uses different logic to find the
bind entry, which fails to find the entry.  This discrepency leads
to a crash.

The fix is to only trim trailing spaces (0x20) from RDN values when
exploding a DN.  The DN explode function is an internal function
when built against the OpenLDAP client libs, but we use an external
MozLDAP call when built against MozLDAP.  This patch makes us always
use our own internal function regardless of the client libs we are
built against.