Bug 604881 - admin server log files have incorrect permissions/ownerships
https://bugzilla.redhat.com/show_bug.cgi?id=604881

Description: Log files generated in the server's startup time has
the server's owner (nsslapd-localuser) for the owner and the group
who starts the server (ordinarily, root).  Once the server goes
into the daemon stage, new log files are owned by owner: nsslapd-
localuser and group: the local user's primary group.  The latter
is the correct ownership of the log files.
This patch sets the correct ownership to the files generated in
the server's startup time, as well.