Bug 676053 - export task followed by import task causes cache assertion
https://bugzilla.redhat.com/show_bug.cgi?id=676053

Description: Task version of export had a bug in handling the busy
instance error case.  When returning due to the busy error, the
function ldbm_back_ldbm2ldif reset the busy bit set by other threads.
This patch checks the special return value set in the busy error
case and resets the busy bit only when it is set by the function.

Also, this patch fixes a bug in dse_add which replaces an adding
entry with its copy in pblock and the original entry is consumed.
But the caller frees the original entry.  Applying this patch,
instead of consuming the original entry, its copy is consumed and
the original adding entry stays in the pblock.