Bug 678646 - Ignore tombstone operations in managed entry plug-in
The managed entry plug-in currently doesn't ignore operations tombstone
entries.  If one is using replication and deletes an origin entry, then
re-adds an origin entry with the same name, the associated managed entry
will be deleted when the tombstone is cleaned up.  This results in an
origin entry with no associated managed entry.

This patch makes the managed entry delete post-op skip processing for
tombstone entries.