Bug 681345 - setup-ds.pl should set SuiteSpotGroup automatically
When the SuiteSpotGroup directive is not specified in the inf file
used by setup, the group permissions do not get set on certain
directories (such as the rundir).  This can break instances if you
are running more than one on the same system.

This patch makes setup use the primary group of the SuiteSpot
user if the group is not specified.  I also added validation of
the group parameter.