Bug 504803 - Allow maxlogsize to be set if logmaxdiskspace is -1
Both the maxlogsize and logmaxdiskspace parameters are allowed to
have values of -1.  If you set logmaxdiskspace to -1 and then
later attempt to set maxlogsize to any other valid value, the
server rejects the change with an operations error.

The problem is that the two parameters are compared to ensure that
maxlogsize is not greater and the logmaxdiskspace.  We need to
skip this check if logmaxdiskspace is unlimited (-1).  I also found
that we were converting -1 to a smaller negative number when doing
the MB->bytes conversion.  This causes other validation errors that
expect -1, but not a smaller negative number.  The fix is to skip
the conversion to bytes and just set a value of -1.