Bug 690882 - (cov#10571) Incorrect sizeof use in uuid code
It looks like the caller wanted to clear out the uuid struct if
there was a problem creating the NSS context. The problem is that
the memset at line 341 is only clearing out the local pointer to
the struct. This local pointer is never seen by the caller and the
struct retains whetever it previously held in memory. We need to
clear out the contents of the struct itself instead of just
clearing the address held by the local pointer.