Bug 690882 - (cov#10572,10710) Incorrect sizeof use in uuid code
When sizeof() is called at line 943, it is calculating the size of
a pointer, not the array. This happens since the array was passed
to get_random_info(), which considers it to just be a pointer. The
result is that we do not fill in the full array with random data.
We should pass the array length to this function instead of using
sizeof().