Bug 690955 - Mrclone fails due to the replica generation id mismatch
https://bugzilla.redhat.com/show_bug.cgi?id=690955

Description: To fix Bug 624442 - MMR: duplicate replica ID,
_replica_configure_ruv (repl5_replica.c) checks the replica IDs
from changelog, backend, and configuration.  If the id from
config is different, recreate RUV.

For the mrclone (restore per backend instance) case, replica ID
from the backend is the other master's one.  The value should
not be used to check if the RUV is recreated or not.  Instead,
it does between the nsds5replicaID from config and the one
from the changelog.