Bug 693466 - Unable to change schema online
https://bugzilla.redhat.com/show_bug.cgi?id=693466
Resolves: bug 693466
Bug Description: Unable to change schema online
Reviewed by: nkinder (Thanks!)
Branch: RHEL-6
Fix Description: The problem is only with attributes that specify a SUP.
There was some old code that assumed the attributes were stored in a list
in order of SUP.  This is not the case.  They are stored in a hashtable
which is not in any order at all.  But the check is redundant anyway.
If a SUP is specified, the code already checks if the superior exists and
returns an error if not.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no
(cherry picked from commit 466fced49dd1b6893c19ce5a7ecd8121f6eceeab)
(cherry picked from commit 5c1cd5f4b3519530ac62df3dd31291e4bf65852a)