Bug 692991 - rhds82 - windows_tot_run: failed to obtain data to send to the consumer; LDAP error - -1
https://bugzilla.redhat.com/show_bug.cgi?id=692991
Resolves: bug 692991
Bug Description: windows_tot_run: failed to obtain data to send to the consumer; LDAP error - -1
Reviewed by: nkinder (Thanks!)
Branch: master
Fix Description: I could not reproduce the problem.  I could not figure out
a way to get Windows to let me add a user-type entry with child entries.
The GUI does not even have this option.  Using ldapmodify results in a
Naming Violation.  I suspect Exchange somehow tweaks AD to allow this
sort of dit structure.  I'm pretty sure the problem is that we are doing
a scope SUBTREE search to get the user's entry and we are running into a
SIZELIMIT_EXCEEDED.  The solution is to just do an LDAP_SCOPE_BASE search
to get the single entry to operate on.  I had to extend
windows_search_entry_ext to add a scope parameter.  Since the old default
was to do a SUBTREE search, I made windows_search_entry use that.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no