Bug 696407 - If an entry with a mixed case RDN is turned to be a tombstone, it fails to assemble DN from entryrdn
https://bugzilla.redhat.com/show_bug.cgi?id=696407

Description: When an entry is turned to be a tombstone, the RDN
part is replaced with this unusual format:
    "nsuniqueid=9a63a581-...-ca8fe9f7,uid=TestUser"
A helper function slapi_rdn_replace_rdn was introduced to support
the replace task, in which function, normalized rdn field "nrdn"
was supposed to have lowered case RDN, but it was not. This patch
fixes it.