Bug 695779 - windows sync can lose old values when a new value is added
When a single modify operation with multiple mods is played back to
AD, it is possible for some old values to be lost.  We check if a
mod is necessary to send to AD to detect loops, and skip sending
operations or values that are not necessary.  This doesn't take
prior mods in the same modify operation into account though, which
can make the results incorrect.

This patch makes the sync code keep a running copy of the AD entry
that the mods are applied to.  We use this copy to check which mods
can be skipped and which we must send.
(cherry picked from commit 2be27d35da9fdf6d5f35ff4fb94906f4fe193133)