Bug 703990 - Support upgrade from Red Hat Directory Server
https://bugzilla.redhat.com/show_bug.cgi?id=703990
Resolves: bug 703990
Bug Description: Support upgrade from Red Hat Directory Server
Reviewed by: nkinder (Thanks!)
When doing an upgrade to a machine of a new architecture, allow the
database to be upgraded "in place" by setup-ds.pl -u with LDIF files.
If there is a file in the database ldif directory called
backendname.upgrade.ldif, this file will be imported and renamed
if the import was successful.
Also fixed a bug in setup-ds.res
Flag Day: yes
Docs: yes