Bug 703304 - Auto membership alternate config area should override default area
The auto membership plug-in should make the alternate config area
(specified as nsslapd-pluginConfigArea) override the default config area.
It currently uses both areas, which could make migrating config from the
default area to a shared/replicated area more difficult.  I don't see a
strong need for having 2 config areas in use at once, so an override
makes sense.

To make the override work, we only need to modify the config loading
code to load the alternate area if one is configured, otherwise we
load from the default area.  The helper function that checks if a DN
is a config entry needs to follow the same logic.