Bug 697961 - memberOf needs to be triggered by internal operations
The memberOf plug-in is not currently triggered by internal modify,
add, delete, or rename operations.  This patch makes it work with
internal operations so it can play nicely with other plug-ins that
make membership changes, such as the new Auto Membership plug-in.

The precedence for the referential integrity plug-in needed to be
lowered to ensure consistency of membership attributes when both
memberOf and referential integrity plug-ins are in use.